requirements.txt
Flask==2.3.3
openai==1.14.3
python-dotenv==1.0.1
captacity